Step 1 to cleanup the IbisManager
