+++ core/src/test/java/org/frankframework/metrics/TestLocalStatisticsRegistry.java
@@ -51,7 +51,7 @@ public class TestLocalStatisticsRegistry {
 		configuration.configure();
 
 		// Ensure only 1 adapter exists
-		assertEquals(1, configuration.getAdapterList().size());
+		assertEquals(1, configuration.getRegisteredAdapters().size());
 	}
 
 	@SuppressWarnings("unchecked")
